Remove ZF dependencies
The ZF components are now Laminas, where console is archived and http is
in security fixes only mode.

Therefore, we extract GetOpt to its own class and write trivially simple
HttpClient and Response classes that are only intended to work with this
script.  This removes dependencies on third-party packages going foward.